requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
opencv-python==4.8.1.78
deepface==0.0.79
Pillow==10.1.0
numpy==1.24.3
dnspython==2.4.2
httpcore==1.0.2
